% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_manhattan.R
\name{fig_manhattan}
\alias{fig_manhattan}
\title{Manhattan plot}
\usage{
fig_manhattan(
  data,
  colours = c("#7395D3", "#1D345D"),
  rank_pos = FALSE,
  thin_thresh = NULL,
  block_thresh = NULL,
  interactive = FALSE,
  interactive_n = NULL,
  thresh = c(1e-05, 5e-08),
  thresh_size = 0.5,
  thresh_colours = c("grey50", "red"),
  trunc = 1e-30,
  highlight_colours = NULL,
  point_size = 2,
  chr_dist = 1e+07,
  x_labels = TRUE,
  label_top = TRUE,
  label_thresh = 1e-05,
  label_size = 3,
  label_ylim = -log10(1e-05),
  label_nudge_y = 0,
  label_box = FALSE,
  title = NULL,
  title_size = NULL,
  title_center = FALSE,
  axis_text_size = NULL,
  axis_title_size = NULL,
  plot_width = 16,
  plot_height = 8,
  girafe = TRUE
)
}
\arguments{
\item{data}{a \code{data.frame} containing the chromosome-position information and
association statistics for each genomic marker (e.g., genetic variants)
with the following columns:
\itemize{
\item{\code{chr}} {
the chromosome for each genomic marker
}
\item{\code{pos}} {
the genomic position for each genomic marker
}
\item{\code{pvalue}} {
the association p-value for each genomic marker
}
\item{\code{highlight}} {
the optional highlight points variable, where
\code{0} = point not highlighted,
\code{1} = first highlight colour,
\code{2} = second highlight colour, etc.
}
\item{\code{highlight_shape}} {
the optional highlight point shape variable, where
\code{0} = standard circle,
\code{1} = standard circle with border,
\code{2} = standard rectangle,
\code{3} = standard rectangle with border,
\code{4} = standard diamond,
\code{5} = standard diamond with border
}
\item{\code{label}} {
the optional point labelling variable (e.g. gene name),
if \code{label = ""} for a point then no
label is presented for that point
}
\item{\code{text}} {
the optional hover text variable for interactive plots to
display further information,
if \code{text = ""} for a point then no hover text is
presented for that point
}
}}

\item{colours}{a \code{vector} of colours used to differentiate chromosomes,
these can either be a separate colour for each chromosome or
a pair of colours (default: \code{c("#7395D3", "#1D345D")})}

\item{rank_pos}{a \code{logical} value whether genomic markers should be
plotted by their rank position (default: \code{FALSE})}

\item{thin_thresh}{a \code{numeric} value representing the minimum p-value
threshold for genomic markers to be displayed (default: \code{NULL})}

\item{block_thresh}{a \code{numeric} value for representing a p-value threshold,
above which genomic markers are represented using blocks of colour
(default: \code{NULL})}

\item{interactive}{a \code{logical} value indicating whether the plot
should be interactive (default: \code{FALSE})}

\item{interactive_n}{a \code{numeric} value indicating the number of top
associated points (max = 100,000) to present before using blocks of colour
to minimise file size (default: \code{NULL})}

\item{thresh}{a \code{numeric} \code{vector} providing p-value thresholds to be plotted
(default: \code{c(1e-5, 5e-8)})}

\item{thresh_size}{a \code{numeric} value indicating the width of the lines
indicating the p-value thresholds (default: \code{0.5})}

\item{thresh_colours}{a \code{character} \code{vector} indicating the colours of the
lines indicating the p-value thresholds (default: \code{c("grey50", "red")})}

\item{trunc}{a \code{numeric} value representing the maximum p-value for which
results are displayed (default: \code{1e-30})}

\item{highlight_colours}{a \code{vector} indicating the colours for the
highlighted genomic markers (default: \code{NULL})}

\item{point_size}{a \code{numeric} value indicating the size of each point
(default: \code{2})}

\item{chr_dist}{a \code{numeric} value indicating the gap between different
chromosomes (default: \code{10000000})}

\item{x_labels}{a \code{logical} value whether the x-axis should be labelled
(default: \code{TRUE})}

\item{label_top}{a \code{logical} value whether the top associated points should
be labelled (default: \code{TRUE})}

\item{label_thresh}{a \code{numeric} value providing a p-value threshold for
labelling points (default: \code{1e-5})}

\item{label_size}{a \code{numeric} value indicating the size of each label
(default: \code{3})}

\item{label_ylim}{a \code{numeric} value indicating maximum y-axis value at which
labels can be displayed (default: \code{-log10(1e-5)})}

\item{label_nudge_y}{a \code{numeric} value indicating the degree to which label
placement on the y-axis should be adjusted (default: \code{0})`}

\item{label_box}{a \code{logical} value indicating whether labels should be
surrounded by a box (default: \code{FALSE})}

\item{title}{a \verb{character`` string providing a title for the plot (default: }NULL`)}

\item{title_size}{a \code{numeric} value indicating the size of the title text
for the plot (default: \code{NULL})}

\item{title_center}{a \code{numeric} value indicating whether the plot title
should be centered (default: \code{FALSE})}

\item{axis_text_size}{a \code{numeric} value indicating the size of the axis
text for the plot (default: \code{NULL})}

\item{axis_title_size}{a \code{numeric} value indicating the size of the axis
title text for the plot (default: \code{NULL})}

\item{plot_width}{a \code{numeric} value indicating the width of the plot
(default: \code{16})}

\item{plot_height}{a \code{numeric} value indicating the height of the plot
(default: \code{8})}

\item{girafe}{a \code{logical} value indicating whether an interactive plot
should be turned into an interactive graphic using
\link[ggiraph:girafe]{girafe()} (default = \code{TRUE})}
}
\value{
\code{fig_manhattan} returns a Manhattan plot for genomic markers
from across the genome, e.g. results from genome-wide association studies.
}
\description{
\code{fig_manhattan} creates a Manhattan plot for genomic markers
from across the genome, e.g. results from genome-wide association studies.
}
\details{
This plotting function plots a Manhattan plot for genomic markers
from across the genome. The default is to truncate these results to
p-value cut-off of \code{1e-30}.
}
\examples{
fig_manhattan(
  data = geni.plots::geni_test_manhattan,
  block_thresh = 1e-4,
  label_box = TRUE
)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
