% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_phewas.R
\name{fig_phewas}
\alias{fig_phewas}
\title{PheWAS plot}
\usage{
fig_phewas(
  data,
  groups = NULL,
  colours = NULL,
  interactive = FALSE,
  thresh = 1e-05,
  thresh_size = 0.5,
  trunc = 1e-30,
  point_size = 2,
  group_dist = 0.05,
  x_labels = TRUE,
  label_top = TRUE,
  label_thresh = 1e-05,
  label_n = NULL,
  label_size = 3,
  label_ymax = 1e-05,
  label_box = FALSE,
  label_nudge_x = 0,
  label_nudge_y = 0,
  title = NULL,
  title_size = NULL,
  title_center = FALSE,
  axis_text_size = NULL,
  axis_text_angle = -60,
  axis_title_size = NULL,
  legend = FALSE,
  legend_title = "Group",
  legend_text_size = NULL,
  legend_title_size = NULL,
  legend_point_size = NULL,
  legend_spacing_size = NULL,
  limit_padding = 20,
  plot_width = 9,
  plot_height = 6,
  girafe = TRUE
)
}
\arguments{
\item{data}{a \code{data.frame} containing the association statistics for each
phenotype with the following columns:
\itemize{
\item{\code{pvalue}} {
the association p-value for each phenotype
}
\item{\code{sign}} {
the direction of the association with the phenotype, where
\code{0} = missing,
\code{1} positive association,
\code{-1} negative association
}
\item{\code{group}} {
the phenotype group for each phenotype
}
\item{\code{label}} {
the optional point labelling variable (e.g. phenotype name),
if \code{label = ""} for a point then no
label is presented for that point
}
\item{\code{text}} {
the optional hover text variable for interactive plots to
display further information,
if \code{text = ""} for a point then no hover text is
presented for that point
}
}}

\item{groups}{a \code{character} \code{vector} of groups describing the grouping
variable in \code{data} (default: \code{NULL})}

\item{colours}{a \code{character} \code{vector} of colours corresponding to defined
groups (default: \code{NULL})}

\item{interactive}{a \code{logical} value indicating whether the plot
should be interactive (default: \code{FALSE})}

\item{thresh}{a \code{numeric} value providing the p-value threshold to be
plotted (default: \code{NULL})}

\item{thresh_size}{a \code{numeric} value indicating the width of the lines
indicating the p-value thresholds (default: \code{0.5})}

\item{trunc}{a \code{numeric} value representing the maximum p-value for which
results are displayed (default: \code{1e-30})}

\item{point_size}{a \code{numeric} value indicating the size of each point
(default: \code{2})}

\item{group_dist}{a \code{numeric} value indicating the gap between different
groups (default: \code{0.05})}

\item{x_labels}{a \code{logical} value whether the x-axis should be labelled
(default: \code{TRUE})}

\item{label_top}{a \code{logical} value whether the top associated points should
be labelled (default: \code{TRUE})}

\item{label_thresh}{a \code{numeric} value providing a p-value threshold for
labelling points (default: \code{1e-5})}

\item{label_n}{an \code{integer} value providing a limit on the number of top
associations to label (default: \code{NULL})}

\item{label_size}{a \code{numeric} value indicating the size of each label
(default: \code{3})}

\item{label_ymax}{a \code{numeric} value indicating the p-value threshold for the
maximum y-axis value at which labels can be displayed (default: \code{1e-5})}

\item{label_box}{a \code{logical} value indicating whether labels should be
surrounded by a box (default: \code{FALSE})}

\item{label_nudge_x}{a \code{numeric} value indicating the degree to which label
placement on the x-axis should be adjusted (default: \code{0})}

\item{label_nudge_y}{a \code{numeric} value indicating the degree to which label
placement on the y-axis should be adjusted (default: \code{0})`}

\item{title}{a \code{character} string providing a title for the plot
(default: \code{NULL})}

\item{title_size}{a \code{numeric} value indicating the size of the title text
for the plot (default: \code{NULL})}

\item{title_center}{a \code{logical} value indicating whether the plot title
should be centered (default: \code{FALSE})}

\item{axis_text_size}{a \code{numeric} value indicating the size of the axis
text for the plot (default: \code{NULL})}

\item{axis_text_angle}{a \code{numeric} value indicating the angle of the text
on the x-axis of the plot (default: \code{-60})`}

\item{axis_title_size}{a \code{numeric} value indicating the size of the axis
title text for the plot (default: \code{NULL})}

\item{legend}{a \code{logical}  value indicating whether a legend corresponding
to the displayed groups should be included (default: \code{FALSE})}

\item{legend_title}{a \code{character} string providing a title for the legend
(default: \code{"Group"})}

\item{legend_text_size}{a \code{numeric} value indicating the size of the legend
text (default: \code{NULL})}

\item{legend_title_size}{a \code{numeric} value indicating the size of the legend
title (default: \code{NULL})}

\item{legend_point_size}{a \code{numeric} value indicating the size of each point
within the legend (default: \code{NULL})}

\item{legend_spacing_size}{a \code{numeric} value indicating spacing of points
present in the legend (default: \code{NULL})}

\item{limit_padding}{a \code{numeric} value indicating the relative distance of
plotted points from x-axis extremes (default: \code{20})}

\item{plot_width}{a \code{numeric} value indicating the width of the PheWAS
plot (default: \code{9})}

\item{plot_height}{a \code{numeric} value indicating the height of the PheWAS
plot (default: \code{6})}

\item{girafe}{a \code{logical} value indicating whether an interactive plot
should be turned into an interactive graphic using
\link[ggiraph:girafe]{girafe()} (default = \code{TRUE})}
}
\value{
\code{fig_phewas} returns a PheWAS plot for phenome-wide association
studies.
}
\description{
\code{fig_phewas} creates a plot visualising results from
phenome-wide association studies (PheWAS).
}
\details{
This plotting function visualises results from phenome-wide
association studies (PheWAS) in the form of a Manhattan style plot.
Associations are grouped into phenotype categories. By default the
results are truncated using a p-value cut-off of \code{1e-30}.
}
\examples{
fig_phewas(
  data = geni.plots::geni_test_phewas
)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
