% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_qq.R
\name{fig_qq}
\alias{fig_qq}
\title{QQ plot}
\usage{
fig_qq(
  pvalues = NULL,
  group = NULL,
  data = NULL,
  groups = NULL,
  colours = NULL,
  interactive = FALSE,
  thresh = NULL,
  sample = FALSE,
  sample_thresh = 0.1,
  sample_prop = 0.1,
  ci = TRUE,
  ci_alpha = 0.05,
  ci_print = FALSE,
  inf_factor = FALSE,
  point_size = 3,
  label_top = FALSE,
  label_thresh = 1e-05,
  label_n = 10,
  label_size = 3.25,
  label_xlim = NULL,
  label_box = FALSE,
  title = NULL,
  title_size = NULL,
  title_center = FALSE,
  axis_text_size = NULL,
  axis_title_size = NULL,
  legend = TRUE,
  legend_title = "Group",
  legend_text_size = NULL,
  legend_title_size = NULL,
  legend_point_size = NULL,
  legend_spacing_size = NULL,
  ymax = NULL,
  plot_width = 6,
  plot_height = 6,
  girafe = TRUE
)
}
\arguments{
\item{pvalues}{the association p-value for each marker
(default: \code{NULL})}

\item{group}{a \code{character} \code{vector} determining the observation group for
each marker (default: \code{NULL})}

\item{data}{a \code{data.frame} containing the association statistics for each
marker with the following columns:
\itemize{
\item{\code{pvalue}} {
the association p-value for each marker
}
\item{\code{group}} {
the optional grouping variable for each marker
}
\item{\code{label}} {
the optional point labelling variable (e.g. genomic marker),
if \code{label = ""} for a point then no
label is presented for that point
}
\item{\code{text}} {
the optional hover text variable for interactive plots to
display further information,
if \code{text = ""} for a point then no hover text is
presented for that point
}
}}

\item{groups}{a \code{character} \code{vector} of groups describing the grouping
variable in \code{data} (default: \code{NULL})}

\item{colours}{a \code{character} \code{vector} of colours corresponding to
defined groups (default: \code{NULL})}

\item{interactive}{a \code{logical} value indicating whether the plot
should be interactive (default: \code{FALSE})}

\item{thresh}{a \code{numeric} value providing the p-value threshold to be
plotted (default: \code{NULL})}

\item{sample}{a \code{logical} value indicating whether a random subset of
p-values above the plotting threshold should be plotted, the number of
which is controlled by \code{sample_prop} (default = \code{FALSE})}

\item{sample_thresh}{a \code{numeric} value indicating the p-value threshold
defining the sample from which additional observations are selected
(default: \code{0.1})}

\item{sample_prop}{a \code{numeric} value indicating the proportion of
sampled observations to be plotted (default: \code{0.1})}

\item{ci}{a \code{logical} value indicating whether confidence intervals
should be displayed (default: \code{TRUE})}

\item{ci_alpha}{a \code{numeric} value providing the threshold defining the
plotted confidence interval (default: \code{0.05})}

\item{ci_print}{a \code{logical} value indicating whether the proportion of points
contained within the confidence interval band should be printed
(default: \code{FALSE})}

\item{inf_factor}{a \code{logical} value indicating whether the inflation factor
should be added to the plot (default: \code{FALSE})}

\item{point_size}{a \code{numeric} value indicating the size of each point
(default: \code{3})}

\item{label_top}{a \code{logical} value whether the top associated points should
be labelled (default: \code{TRUE})}

\item{label_thresh}{a \code{numeric} value providing a p-value threshold for
labelling points (default: \code{1e-5})}

\item{label_n}{an \code{integer} value providing a limit on the number of top
associations to label (default: \code{NULL})}

\item{label_size}{a \code{numeric} value indicating the size of each label
(default: \code{3})}

\item{label_xlim}{a \code{numeric} value indicating maximum x-axis value at
which labels can be displayed (default: \code{NULL})}

\item{label_box}{a \code{logical} value indicating whether labels should be
surrounded by a box (default: \code{FALSE})}

\item{title}{a \code{character} string providing a title for the plot
(default: \code{NULL})}

\item{title_size}{a \code{numeric} value indicating the size of the title text
for the plot (default: \code{NULL})}

\item{title_center}{a \code{logical} value indicating whether the plot title
should be centered (default: \code{FALSE})}

\item{axis_text_size}{a \code{numeric} value indicating the size of the axis
text for the plot (default: \code{NULL})}

\item{axis_title_size}{a \code{numeric} value indicating the size of the axis
title text for the plot (default: \code{NULL})}

\item{legend}{a \code{logical}  value indicating whether a legend corresponding
to the displayed groups should be included (default: \code{FALSE})}

\item{legend_title}{a \code{character} string providing a title for the legend
(default: \code{"Group"})}

\item{legend_text_size}{a \code{numeric} value indicating the size of the legend
text (default: \code{NULL})}

\item{legend_title_size}{a \code{numeric} value indicating the size of the legend
title (default: \code{NULL})}

\item{legend_point_size}{a \code{numeric} value indicating the size of each point
within the legend (default: \code{NULL})}

\item{legend_spacing_size}{a \code{numeric} value indicating spacing of points
present in the legend (default: \code{NULL})}

\item{ymax}{a \code{numeric} value defining the maximum value of the y-axis
(default: \code{NULL})}

\item{plot_width}{a \code{numeric} value indicating the width of the plot
(default: \code{6})}

\item{plot_height}{a \code{numeric} value indicating the height of the plot
(default: \code{6})}

\item{girafe}{a \code{logical} value indicating whether an interactive plot
should be turned into an interactive graphic using
\link[ggiraph:girafe]{girafe()} (default = \code{TRUE})}
}
\value{
\code{fig_qq} returns a quantile-quantile plot.
}
\description{
\code{fig_qq} creates a quantile-quantile (QQ) plot.
}
\details{
This plotting function plots a quantile-quantile plot of
-log10(p-values). Observations can be divided into groups and can include
corresponding confidence intervals. This plot is based on the
\href{https://genome.sph.umich.edu/wiki/Code_Sample:_Generating_QQ_Plots_in_R}{QQ plot}
by Matthew Flickinger.
}
\examples{
fig_qq(
  pvalues = geni.plots::geni_test_phewas$pvalue
)
fig_qq(
  data = geni.plots::geni_test_phewas \%>\%
    select(pvalue, group, label, text),
  interactive = TRUE
)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
