% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_region.R
\name{fig_region_save}
\alias{fig_region_save}
\title{Regional plot save}
\usage{
fig_region_save(
  fig,
  file,
  ntraits = 1,
  interactive = FALSE,
  width = NULL,
  height = NULL,
  dpi = 500
)
}
\arguments{
\item{fig}{a regional plot object}

\item{file}{a \code{character} string of the file path to save the plot}

\item{ntraits}{an \code{integer} value of the number of traits plotted
(default: \code{1})}

\item{interactive}{a \code{logical} value indicating whether the plot
is interactive (default: \code{FALSE})}

\item{width}{a \code{numeric} value indicating the width of the plot
in inches (default: \code{NULL})}

\item{height}{a \code{numeric} value indicating the height of the plot
in inches (default: \code{NULL})}

\item{dpi}{the resolution of the plot (default: \code{500})}
}
\description{
\code{fig_region_save} saves the regional plots to file.
}
\examples{
\dontrun{
fig <- fig_region(
  data = geni.plots::geni_test_region$assoc,
  corr = geni.plots::geni_test_region$corr,
  build = 37
)
fig_region_save(fig, "test.png")
}
}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
