% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_region.R
\name{fig_region_stack}
\alias{fig_region_stack}
\title{Stacked regional plot}
\usage{
fig_region_stack(
  data,
  traits,
  corr = NULL,
  corr_top = NULL,
  top_marker = NULL,
  r2 = FALSE,
  build = 38,
  prob = FALSE,
  interactive = FALSE,
  thresh = NULL,
  thresh_colour = "grey50",
  x_min = NULL,
  x_max = NULL,
  y_title = NULL,
  point_size = 3,
  alpha = 1,
  genebar = TRUE,
  genebar_ntracks = NULL,
  genebar_label_pos = 3.6,
  genebar_label_size = 4,
  genebar_line_size = 0.8,
  label_size = 3.5,
  highlights = NULL,
  highlights_cat = NULL,
  highlights_label = TRUE,
  highlights_shape = 22,
  highlights_nolabel_shape = 21,
  highlights_sort = TRUE,
  highlights_colours = NULL,
  highlights_title = "Group",
  title_size = 16,
  title_center = FALSE,
  axis_text_size = 14,
  axis_title_size = 16,
  legend = TRUE,
  legend_text_size = 12,
  legend_title_size = 12,
  point_padding = 0,
  nudge_x = 0,
  nudge_y = 0,
  nudge_y_top = 0.06,
  ylim_prob = 1,
  assoc_plot_size = NULL,
  genebar_plot_size = NULL,
  legend_plot_dist = NULL,
  plot_width = NULL,
  plot_height = NULL,
  girafe = TRUE
)
}
\arguments{
\item{data}{a \code{data.frame} containing the association statistics for each
marker within a genomic region with the following columns:
\itemize{
\item{\code{marker}} {
the genomic marker identifier (e.g. rsID)
}
\item{\code{chr}} {
the chromosome for each genomic marker
}
\item{\code{pos}} {
the genomic position for each genomic marker
}
and one of the following:
\item{\code{pvalue_1}, \code{pvalue_2}, \code{pvalue_3}, etc.} {
the association p-values for each genomic marker and trait
}
\item{\code{z_1}, \code{z_2}, \code{z_3}, etc.} {
the association z-statistics for each genomic marker and trait
}
\item{\code{prob_1}, \code{prob_2}, \code{prob_3}, etc.} {
the association probabilities for each genomic marker and trait
}
}}

\item{traits}{a \code{character} \code{vector} of trait names}

\item{corr}{a \code{numeric} \code{matrix} of correlation statistics between
the markers (default: \code{NULL})}

\item{corr_top}{a \code{numeric} \code{vector} of correlation statistics between the
top marker and the rest of the markers (default: \code{NULL})}

\item{top_marker}{a \code{character} value depicting the marker to plot the
correlation statistics of the rest of the markers against (default: \code{NULL})}

\item{r2}{a \code{logical} value indicating whether the set of correlation
statistics entered in \code{corr} or \code{corr_top} are squared (default: \code{FALSE})}

\item{build}{a \code{numeric} value indicating the genome build used to determine
genomic position (default: \code{38} representing human assembly GRCh38)}

\item{prob}{a \code{logical} value indicating whether probability statistics
should be plotted instead of -log10(p-values) (default: \code{FALSE})}

\item{interactive}{a \code{logical} value indicating whether the plot
should be interactive (default: \code{FALSE})}

\item{thresh}{a \code{numeric} \code{vector} providing the p-value thresholds to be
plotted (default: \code{NULL})}

\item{thresh_colour}{a \code{character} \code{vector} indicating the colours of the
lines indicating the p-value thresholds (default: \code{"grey50"})}

\item{x_min}{a \code{numeric} value depicting the minimum plotted x-axis value
representing the start of the genomic region (default: \code{NULL})}

\item{x_max}{a \code{numeric} value depicting the maximum plotted x-axis value
representing the end of the genomic region (default: \code{NULL})}

\item{y_title}{a \code{character} string defining the title of the y-axis
(default: \code{NULL})}

\item{point_size}{a \code{numeric} value indicating the size of each point
(default: \code{3})}

\item{alpha}{a \code{numeric} value adjusting the opacity of colours
representing the correlation statistics (default: \code{1})}

\item{genebar}{a \code{logical} value indicating whether bars representing
the genes should be included in the plot (default: \code{TRUE})}

\item{genebar_ntracks}{an \code{integer} value indicating the number of
tracks to be included in the gene bar (default: \code{NULL})}

\item{genebar_label_pos}{a \code{numeric} value indicating the relative
position of gene labels with respect to each gene bar (default: \code{3.6})}

\item{genebar_label_size}{a \code{numeric} value defining the size of each
gene label (default: \code{4.25})}

\item{genebar_line_size}{a \code{numeric} value defining the line size of
each gene bar (default: \code{0.8})}

\item{label_size}{a \code{numeric} value indicating the size of each label
(default: \code{3.5})}

\item{highlights}{a \code{character} \code{vector} defining a set of markers to
highlight in the plot (default: \code{NULL})}

\item{highlights_cat}{a \code{character} \code{vector} defining the category for
each highlighted marker (default: \code{NULL})}

\item{highlights_label}{a \code{logical} value indicating whether highlighted
points should be labelled (default: \code{TRUE})}

\item{highlights_shape}{a value defining the shape for highlighted points
(default: \code{22})}

\item{highlights_nolabel_shape}{a value defining the shape for points
which are not highlighted (default: \code{21})}

\item{highlights_sort}{a \code{logical} value indicating whether to sort highlight
group label levels (default: \code{TRUE})}

\item{highlights_colours}{a \code{character} \code{vector} specifying colours for
highlighted points (default: \code{NULL})}

\item{highlights_title}{a \code{character} string providing a title for the legend
corresponding to the highlighted points (default: \code{"Group"})}

\item{title_size}{a \code{numeric} value indicating the size of the title text
for the plot (default: \code{NULL})}

\item{title_center}{a \code{logical} value indicating whether the plot title
should be centered (default: \code{FALSE})}

\item{axis_text_size}{a \code{numeric} value indicating the size of the axis
text for the plot (default: \code{14})}

\item{axis_title_size}{a \code{numeric} value indicating the size of the axis
title text for the plot (default: \code{16})}

\item{legend}{a \code{logical}  value indicating whether a legend corresponding
to the displayed groups should be included (default: \code{TRUE})}

\item{legend_text_size}{a \code{numeric} value indicating the size of the legend
text (default: \code{NULL})}

\item{legend_title_size}{a \code{numeric} value indicating the size of the legend
title (default: \code{NULL})}

\item{point_padding}{a \code{numeric} value indicating the relative distance of
labels from plotted points (default: \code{0}}

\item{nudge_x}{a \code{numeric} value indicating the degree to which label
placement on the x-axis should be adjusted (default: \code{0})}

\item{nudge_y}{a \code{numeric} value indicating the degree to which label
placement on the y-axis should be adjusted (default: \code{0})}

\item{nudge_y_top}{a \code{numeric} value indicating the degree to which the top
marker should be adjusted on the y-axis by a proportion of the y-axis limit
(default: \code{0.06})}

\item{ylim_prob}{a \code{numeric} value defining the upper y-axis limit for
probability plots (default: \code{1})}

\item{assoc_plot_size}{a \code{numeric} value determining the size of the
association plot (default: \code{NULL})}

\item{genebar_plot_size}{a \code{numeric} value determining the size of the gene
bar plot (default: \code{NULL})}

\item{legend_plot_dist}{a \code{numeric} value defining the distance and size of
the legend from the bottom of the regional plot (default = \code{NULL})}

\item{plot_width}{a \code{numeric} value indicating the width of the plot
(default: \code{NULL})}

\item{plot_height}{a \code{numeric} value indicating the height of the plot
(default: \code{NULL})}

\item{girafe}{a \code{logical} value indicating whether an interactive plot
should be turned into an interactive graphic using
\link[ggiraph:girafe]{girafe()} (default = \code{TRUE})}
}
\value{
\code{fig_region_stack} returns a stacked regional plot.
}
\description{
\code{fig_region_stack} creates a stacked regional association plot.
}
\examples{
fig <- fig_region_stack(
  data = geni.plots::geni_test_stack_region$assoc,
  traits = c("Interleukin-6 levels", "Interleukin-6 receptor levels"),
  corr = geni.plots::geni_test_stack_region$corr,
  build = 37,
  highlights = "rs11265611",
  title_center = TRUE
)

# Notes:
#   (i) corr has to have the same markers as assoc in the same order
#   (ii) by default fig_region_stack assumes corr contains correlation
#        statistics that have not been squared

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}
}
