% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geni.plots-package.R
\docType{package}
\name{geni.plots-package}
\alias{geni.plots}
\alias{geni.plots-package}
\title{geni.plots: GENI plots}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

GENI plots is designed to visualise results from genome-wide association studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrs95/geni.plots}
  \item \url{https://jrs95.github.io/geni.plots/}
  \item Report bugs at \url{https://github.com/jrs95/geni.plots/issues}
}

}
\author{
\strong{Maintainer}: James Staley \email{jrstaley95@gmail.com} (\href{https://orcid.org/0009-0001-9520-5011}{ORCID})

Authors:
\itemize{
  \item Wes Spiller (\href{https://orcid.org/0000-0002-8169-5531}{ORCID})
}

}
\keyword{internal}
