% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geni_test.R
\docType{data}
\name{geni_test_manhattan}
\alias{geni_test_manhattan}
\title{GENI plots test data for the Manhattan plot}
\format{
A \code{data.frame} containing the variables necessary for
constructing a static Manhattan plot. Specifically, this includes
the following information:

\itemize{
\item{\code{chr}} {
the chromosome for each genomic marker
}
\item{\code{pos}} {
the genomic position (GRCh37) for each genomic marker
}
\item{\code{pvalue}} {
the association p-value for each genomic marker
}
\item{\code{highlight}} {
an indicator variable whether the genomic marker should be highlighted:
\code{0} = point not highlighted,
\code{1} = first highlight colour
}
\item{\code{highlight_shape}} {
an indicator variable of the illustrated shape for the genomic markers:
\code{0} = standard circle,
\code{5} = standard diamond with border
}
\item{\code{label}} {
the name of the nearest gene for each sentinel genome-wide
associated genetic variant, the labels for non-sentinel
genetic variants are omitted
}
}
}
\usage{
geni_test_manhattan
}
\description{
An R object containing formatted results from a genome-wide
association study (GWAS) of interleukin-6 levels with \emph{p}-values < \code{1e-3}
(\href{https://www.ebi.ac.uk/gwas/studies/GCST90274815}{GCST90274815}).
}
\examples{
head(geni_test_manhattan)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}

Wes Spiller
}
\keyword{datasets}
