% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geni_test.R
\docType{data}
\name{geni_test_phewas}
\alias{geni_test_phewas}
\title{GENI plots test data for the PheWAS plot}
\format{
A \code{data.frame} containing the variables necessary for
constructing an interactive PheWAS plot. Specifically, this includes
the following information:

\itemize{
\item{\code{pvalue}} {
the association p-value for each phenotype
}
\item{\code{sign}} {
the direction of the association with the phenotype, where
\code{0} = missing,
\code{1} positive association,
\code{-1} negative association
}
\item{\code{group}} {
the phenotype group for each phenotype
}
\item{\code{label}} {
the phenotype name
}
\item{\code{text}} {
the hover text containing further information on the association
including: phenotype, phenotypic category, genetic variant ID,
direction of association, association p-value, number of cases
and number of controls
}
}
}
\usage{
geni_test_phewas
}
\description{
An R object containing formatted data from a phenome-wide
association study (PheWAS) of rs2228145
(\href{https://r9.finngen.fi/variant/1:154454494-A-C}{1:154426970-A-C})
from FinnGen r9.
}
\examples{
head(geni_test_phewas)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}

Wes Spiller
}
\keyword{datasets}
