% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geni_test.R
\docType{data}
\name{geni_test_stack_region}
\alias{geni_test_stack_region}
\title{GENI plots test data for the stacked regional plot}
\format{
A \code{list} containing 2 objects with the information necessary
for constructing an interactive stacked regional plot.
Specifically, this includes the following information:

\itemize{
\item{\code{assoc}} {
a \code{data.frame} with genetic association results for interleukin-6 levels
from \href{https://www.ebi.ac.uk/gwas/studies/GCST90274815}{GCST90274815} and
for interleukin-6 receptor levels
\href{https://www.ebi.ac.uk/gwas/studies/GCST90088597}{GCST90088597}
for the region 1:154301970-154551970
}
\item{\code{corr}} {
a \code{matrix} of Pearson correlation statistics (not squared) from the
European samples of
\href{https://www.internationalgenome.org/}{1000 Genomes Phase 3}
for the region 1:154301970-154551970, this \code{matrix} has the same markers
in the same order as \code{assoc}
}
}
}
\usage{
geni_test_stack_region
}
\description{
An R object containing formatted results from genome-wide
association studies (GWAS) of interleukin-6 levels
(\href{https://www.ebi.ac.uk/gwas/studies/GCST90274815}{GCST90274815}) and
interleukin-6 receptor levels
(\href{https://www.ebi.ac.uk/gwas/studies/GCST90088597}{GCST90088597})
and linkage disequilibrium statistics from
\href{https://www.internationalgenome.org/}{1000 Genomes Phase 3}
for the region 1:154301970-154551970.
}
\details{
\code{assoc} is a \code{data.frame} with the following columns:
\itemize{
\item{\code{marker}} {
the genomic marker identifier (i.e. rsID or chromosome-position)
}
\item{\code{chr}} {
the chromosome for each genomic marker
}
\item{\code{pos}} {
the genomic position (GRCh37) for each genomic marker
}
\item{\code{pvalue_1}} {
the association p-value for interleukin-6 levels for each genomic
marker
}
\item{\code{pvalue_2}} {
the association p-value for interleukin-6 receptor levels for each
genomic marker
}
}
}
\examples{
head(geni_test_stack_region$assoc)
head(geni_test_stack_region$corr)

}
\author{
James Staley \href{mailto:jrstaley95@gmail.com}{jrstaley95@gmail.com}

Wes Spiller
}
\keyword{datasets}
